test(e2e): 重写站点看板收藏用例，移除过时卡片版 (#72)
看板重构(#59)后旧卡片 UI（.site-card/.collapse-healthy-toggle 等）已移除；
用 .site-row/.fav/.sname 重写「收藏置顶 + 持久化」，删除已无意义的「折叠健康站点」测试
（展开行为由 sites-board-fixes.spec.ts 覆盖）。
